Separate SH used to run Configure from SH used on target system
Android, for reasons best known to itself, doesn't have a "/bin/sh" it has a "/system/bin/sh".
We need to run "/bin/sh" as $sh inside ./Configure on the host, but store a different path
for use on the actual Android system (eg by backticks, system etc, in SH_PATH).

The new variable is "targetsh" and defaults to "sh" unless changed by the hints file/Configure params.